dev/webapp/requirements-dev.txt
asgiref==3.3.4
azure-cognitiveservices-vision-computervision==0.8.0
azure-common==1.1.27
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
defusedxml==0.6.0
dill==0.3.3
Django==3.2.3
django-allauth==0.42.0
django-appconf==1.0.4
django-background-task==0.1.8
django-bootstrap-form==3.4
django-compressor==2.4.1
django-compressor-toolkit==0.6.0
django-debug-permissions==1.0.0
django-debug-toolbar==2.2
django-decorator-include==3.0
django-formtools==2.3
django-jazzmin==2.4.6
django-json-widget==1.1.1
django-libsass==0.8
django-phonenumber-field==4.0.0
django-sass-processor==1.0.1
django-tables2==2.3.1
djangorestframework==3.12.4
dnspython==2.1.0
et-xmlfile==1.0.1
Faker==0.9.1
filelock==3.0.12
future==0.18.2
greenlet==1.0.0
idna==2.10
isodate==0.6.0
jdcal==1.4.1
joblib==1.0.1
kafka-python==2.0.2
kiwisolver==1.3.1
kq==2.0.1
libsass==0.20.1
matplotlib==3.4.1
mixer==6.1.3
msrest==0.6.21
nltk==3.6.2
numpy==1.19.1
oauthlib==3.1.0
openpyxl==3.0.4
pandas==1.0.5
parameterized==0.7.4
pdf2image==1.13.1
pdfminer.six==20200104
pdfplumber==0.5.22
phonenumbers==8.12.7
Pillow==7.2.0
psycopg2==2.8.6
py3-validate-email==0.2.12
pycparser==2.20
pycryptodome==3.9.8
pyparsing==2.4.7
pytesseract==0.3.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-magic==0.4.18
python3-openid==3.2.0
pytz==2020.1
rcssmin==1.0.6
regex==2021.4.4
requests==2.24.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.15.0
sortedcontainers==2.2.2
SQLAlchemy==1.4.13
sqlparse==0.3.1
text-unidecode==1.2
tqdm==4.60.0
unicodecsv==0.14.1
urllib3==1.25.10
Wand==0.6.2
XlsxWriter==1.2.9
xlwt==1.3.0
